% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/kohonenplot.R
\name{som_to_dataframe}
\alias{som_to_dataframe}
\title{Convert a SOM to a data frame ready to be plotted}
\usage{
som_to_dataframe(som)
}
\arguments{
\item{som}{A kohonen Self-Organising Map}
}
\description{
Takes a SOM and create a data frame with a row for each tile in the SOM and
a column for each column of the original matrix. The content of the data
frame is equivalent to the codes matrix.
}

